Update README.md
In this example, we start by integrating the Coinbase API code that we wrote earlier to retrieve historical Bitcoin price data and save it as a CSV file. We then load the Bitcoin price data into a pandas DataFrame and split it into training and testing sets. We train a linear regression model on the training data and use the trained model to make predictions on the test data. Finally, we evaluate the performance of the model using the mean squared error metric.

Please note that you will need to replace <your_api_key> in the headers dictionary with your actual Coinbase API key for this code to work. Also, you may need to modify the API request parameters in the params dictionary to retrieve the specific Bitcoin price data that you need.